<FIX> person id 자료형 Long으로 변경

+++ src/main/java/roomescape/entity/Person.java
@@ -1,10 +1,10 @@
 package roomescape.entity;
 
 public class Person {
-    private int id;
+    private Long id;
     private String name;
 
-    public Person(int id, String name) {
+    public Person(Long id, String name) {
         this.id = id;
         this.name = name;
     }
